Move favicon functions: basics.js -> favicon.js
